fix: Resolve async operations, circular dependencies, and code organization issues
- Replace std::fs with tokio::fs in auth setup functions for proper async file operations
- Create shared types.rs module to break circular dependency between email core and testing modules
- Make generate_key function synchronous and move use statements to file tops per Rust conventions
- Update main.rs to handle synchronous generate_key function call
- Maintain backward compatibility through proper module re-exports
