+++ crates/cli/src/commands/email.rs
@@ -5,6 +5,7 @@ pub mod providers;
 pub mod queue;
 pub mod templates;
 pub mod testing;
+pub mod types;
 
 // Re-export all public items
 pub use analytics::*;
@@ -13,3 +14,4 @@ pub use providers::*;
 pub use queue::*;
 pub use templates::*;
 pub use testing::*;
+pub use types::*;
